md5: Select backend by cfg-if
