+++ md5/Cargo.toml
@@ -16,6 +16,7 @@ name = "md5"
 
 [dependencies]
 digest = "0.10.7"
+cfg-if = "1.0"
 
 [target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
 md5-asm = { version = "0.5", optional = true }
@@ -29,3 +30,4 @@ default = ["std"]
 std = ["digest/std"]
 asm = ["md5-asm"] # WARNING: this feature SHOULD NOT be enabled by library crates
 oid = ["digest/oid"] # Enable OID support. WARNING: Bumps MSRV to 1.57
+force-soft = [] # Force software implementation
