+++ md5/src/lib.rs
@@ -30,14 +30,9 @@
 )]
 #![warn(missing_docs, rust_2018_idioms)]
 
-#[cfg(all(feature = "asm", any(target_arch = "x86", target_arch = "x86_64")))]
-extern crate md5_asm as compress;
-
-#[cfg(not(all(feature = "asm", any(target_arch = "x86", target_arch = "x86_64"))))]
-mod compress;
-
 pub use digest::{self, Digest};
 
+mod compress;
 use compress::compress;
 
 use core::{fmt, slice::from_ref};
